-Changed Data Stream Communication between services was achieved via REST protocols
