+++ Services/PhoneBookApi/PhoneBookApp.PhoneBookApi/PhoneBookApp.PhoneBookApi.csproj
@@ -20,10 +20,4 @@
     <ProjectReference Include="..\..\..\Shared\PhoneBookApp.Shared\PhoneBookApp.Shared.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Compile Update="Consumer\CreateReportEventConsumer.cs">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-    </Compile>
-  </ItemGroup>
-
 </Project>
