+++ Services/PhoneBookApi/PhoneBookApp.PhoneBookApi/appsettings.json
@@ -1,5 +1,5 @@
 {
-  "RabbitMQUrl": "localhost",
+
   "DatabaseSettings": {
     "ContactCollectionName": "Contacts",
     "ContactInfoCollectionName": "ContactInfos",
